+++ CHANGELOG.md
@@ -2,5 +2,13 @@
 
 All notable changes to this project will be documented in this file, in reverse chronological order by release.
 
-## 0.1.0 - Unreleased
+## 0.3.0 - unreleased
+
+### Added
+
+- Support for ip address geolocation fallback for all providers.
+
+## 0.2.0 - 2017-04-16
+
+Initial release with support for [Combain CPS API](https://combain.com/api/), [Google Maps Geolocation API](https://developers.google.com/maps/documentation/geolocation/intro), [Mozilla Location Service (MLS)](https://location.services.mozilla.com/) and [Unwired Labs LocationAPI](https://unwiredlabs.com/locationapi).